requirements.txt
transformers
sentencepiece
torch
Pillow
torch
accelerate
torchvision
datasets==2.18.0
